requirements.txt
certifi==2019.11.28
chardet==3.0.4
idna==2.8
pendulum==2.0.5
python-dateutil==2.8.1
pytzdata==2019.3
requests==2.22.0
six==1.14.0
urllib3==1.25.8
